(PC-26074)[API] refactor: remove abstract token dependency to user ids
and move methods to follow OOP conventional order :
1. public static methods
2. abstract static methods
3. instance methods

+++ api/src/pcapi/core/token/__init__.py
@@ -7,7 +7,7 @@ import json
 import logging
 import secrets
 import typing
-from typing import Any
+import uuid
 
 from flask import current_app as app
 import jwt
@@ -28,10 +28,13 @@ class TokenType(enum.Enum):
     RESET_PASSWORD = "reset_password"
 
 
+T = typing.TypeVar("T", bound="AbstractToken")
+
+
 @dataclasses.dataclass(frozen=True)
 class AbstractToken(abc.ABC):
     type_: TokenType
-    user_id: int
+    key_suffix: int | str | None
     encoded_token: str
     data: dict
 
@@ -42,47 +45,55 @@ class AbstractToken(abc.ABC):
         CREATE = "create"
 
     @classmethod
-    @abc.abstractmethod
-    def load_without_checking(cls, encoded_token: str, *args: typing.Any, **kwargs: typing.Any) -> "AbstractToken":
-        raise NotImplementedError()
-
-    @classmethod
     def load_and_check(
-        cls,
+        cls: typing.Type[T],
         encoded_token: str,
         type_: TokenType,
         user_id: int | None = None,
-    ) -> "AbstractToken":
+    ) -> T:
         token = cls.load_without_checking(encoded_token, type_=type_, user_id=user_id)
         token.check(type_, user_id)
         return token
 
     @classmethod
-    @abc.abstractmethod
-    def create(cls, type_: TokenType, ttl: timedelta | None, user_id: int, data: dict | None = None) -> "AbstractToken":
-        raise NotImplementedError()
+    def get_redis_key(cls: typing.Type[T], type_: TokenType, key_suffix: int | str | None) -> str:
+        return f"pcapi:token:{type_.value}_{key_suffix}"
 
     @classmethod
-    def _get_redis_key(cls, type_: TokenType, user_id: int) -> str:
-        return f"pcapi:token:{type_.value}_{user_id}"
+    def token_exists(cls: typing.Type[T], type_: TokenType, key_suffix: int | str | None) -> bool:
+        return bool(app.redis_client.exists(cls.get_redis_key(type_, key_suffix)))
 
     @classmethod
-    def get_expiration_date(cls, type_: TokenType, user_id: int) -> datetime | None:
-        key = Token._get_redis_key(type_, user_id)
+    def get_expiration_date(cls: typing.Type[T], type_: TokenType, key_suffix: int | str | None) -> datetime | None:
+        key = cls.get_redis_key(type_, key_suffix)
         ttl = app.redis_client.ttl(key)
         if ttl < 0:
             # -2 if doesn't exist, -1 if no expiration
             return None
         return datetime.utcnow() + timedelta(seconds=ttl)
 
-    def get_expiration_date_from_token(self) -> datetime | None:
-        return self.get_expiration_date(self.type_, self.user_id)
+    @classmethod
+    def delete(cls: typing.Type[T], type_: TokenType, key_suffix: int | str | None) -> None:
+        app.redis_client.delete(cls.get_redis_key(type_, key_suffix))
 
-    def check(self, type_: TokenType, user_id: int | None = None) -> None:
-        redis_key = Token._get_redis_key(self.type_, self.user_id)
+    @classmethod
+    @abc.abstractmethod
+    def load_without_checking(cls: typing.Type[T], encoded_token: str, *args: typing.Any, **kwargs: typing.Any) -> T:
+        raise NotImplementedError()
+
+    @classmethod
+    @abc.abstractmethod
+    def create(cls: typing.Type[T], type_: TokenType, *args: typing.Any, **kwargs: typing.Any) -> T:
+        raise NotImplementedError()
+
+    def get_expiration_date_from_token(self) -> datetime | None:
+        return self.get_expiration_date(self.type_, self.key_suffix)
+
+    def check(self, type_: TokenType, key_suffix: int | str | None = None) -> None:
+        redis_key = AbstractToken.get_redis_key(self.type_, self.key_suffix)
         if (
             self.type_ != type_
-            or (user_id is not None and self.user_id != user_id)
+            or (key_suffix is not None and self.key_suffix != key_suffix)
             or app.redis_client.get(redis_key) != self.encoded_token
         ):
             self._log(self._TokenAction.CHECK_KO)
@@ -90,65 +101,55 @@ class AbstractToken(abc.ABC):
         self._log(self._TokenAction.CHECK_OK)
 
     def expire(self) -> None:
-        app.redis_client.delete(Token._get_redis_key(self.type_, self.user_id))
+        app.redis_client.delete(AbstractToken.get_redis_key(self.type_, self.key_suffix))
         self._log(self._TokenAction.EXPIRE)
 
-    @classmethod
-    def token_exists(cls, type_: TokenType, user_id: int) -> bool:
-        return bool(app.redis_client.exists(cls._get_redis_key(type_, user_id)))
-
     def _log(self, action: _TokenAction) -> None:
-        logger.info("[TOKEN](%s){%i, %s, %s}", action.value, self.user_id, self.type_.value, self.encoded_token)
-
-    @classmethod
-    def delete(cls, type_: TokenType, user_id: int) -> None:
-        app.redis_client.delete(cls._get_redis_key(type_, user_id))
-
-    @classmethod
-    def get_token(cls, type_: TokenType, user_id: int) -> "AbstractToken | None":
-        raise NotImplementedError()
+        logger.info("[TOKEN](%s)%s, %s, %s", action.value, self.key_suffix, self.type_.value, self.encoded_token)
 
 
 class Token(AbstractToken):
+    @property
+    def user_id(self) -> int:
+        assert self.key_suffix
+        return int(self.key_suffix)
+
     @classmethod
     def load_without_checking(cls, encoded_token: str, *args: typing.Any, **kwargs: typing.Any) -> "Token":
         try:
             payload = utils.decode_jwt_token(encoded_token)
-        except jwt.exceptions.ExpiredSignatureError:
-            raise users_exceptions.ExpiredToken()
-        except jwt.PyJWTError:
-            raise users_exceptions.InvalidToken()
-        try:
-            data = payload["data"] if payload["data"] is not None else {}
             type_ = TokenType(payload["token_type"])
             user_id = payload["user_id"]
-            return cls(type_, user_id, encoded_token, data)
-        except (KeyError, ValueError):
-            raise users_exceptions.InvalidToken()
+        except jwt.exceptions.ExpiredSignatureError as e:
+            raise users_exceptions.ExpiredToken() from e
+        except (KeyError, ValueError, jwt.PyJWTError) as e:
+            raise users_exceptions.InvalidToken() from e
+
+        data = payload.get("data", {})
+        return cls(type_, user_id, encoded_token, data)
 
     @classmethod
-    def create(cls, type_: TokenType, ttl: timedelta | None, user_id: int, data: dict | None = None) -> "Token":
-        payload: dict[str, Any] = {
+    def create(
+        cls,
+        type_: TokenType,
+        ttl: timedelta | None,
+        user_id: int,
+        data: dict | None = None,
+    ) -> "Token":
+        payload: dict[str, typing.Any] = {
             "token_type": type_.value,
             "user_id": user_id,
-            "data": data,
+            "data": data or {},
         }
         if ttl:
             payload["exp"] = (datetime.utcnow() + ttl).timestamp()
 
         encoded_token = utils.encode_jwt_payload(payload)
         if ttl is None or ttl > timedelta(0):
-            app.redis_client.set(cls._get_redis_key(type_, user_id), encoded_token, ex=ttl)
+            app.redis_client.set(cls.get_redis_key(type_, user_id), encoded_token, ex=ttl)
         token = Token.load_without_checking(encoded_token)
         token._log(cls._TokenAction.CREATE)
         return token
-
-    @classmethod
-    def get_token(cls, type_: TokenType, user_id: int) -> "Token | None":
-        encoded_token = app.redis_client.get(cls._get_redis_key(type_, user_id))
-        if encoded_token is None:
-            return None
-        return cls.load_without_checking(encoded_token)
 
 
 @dataclasses.dataclass(frozen=True)
@@ -163,48 +164,38 @@ class SixDigitsToken(AbstractToken):
         encoded_token: str,
         type_: TokenType,
         user_id: int,
-        *args: typing.Any,
-        **kwargs: typing.Any,
     ) -> "SixDigitsToken":
+        if user_id is None:
+            raise ValueError("user_id is required for SixDigitsToken")
         try:
             data_json = app.redis_client.get(cls._get_redis_extra_data_key(type_=type_, user_id=user_id))
             if data_json is None:
                 raise users_exceptions.InvalidToken()
             data = json.loads(data_json)
-        except json.JSONDecodeError:
-            raise users_exceptions.InvalidToken()
+        except json.JSONDecodeError as e:
+            raise users_exceptions.InvalidToken() from e
         return cls(type_, user_id, encoded_token, data)
-
-    @classmethod
-    def load_and_check(cls, encoded_token: str, type_: TokenType, user_id: int | None = None) -> "AbstractToken":
-        if user_id is None:
-            raise ValueError("user_id is required for SixDigitsToken")
-        return super().load_and_check(encoded_token, type_, user_id)
 
     @classmethod
     def create(
         cls, type_: TokenType, ttl: timedelta | None, user_id: int, data: dict | None = None
     ) -> "SixDigitsToken":
         encoded_token = "{:06}".format(secrets.randbelow(1_000_000))  # 6 digits
-        app.redis_client.set(cls._get_redis_key(type_, user_id), encoded_token, ex=ttl)
+        app.redis_client.set(cls.get_redis_key(type_, user_id), encoded_token, ex=ttl)
         json_data = json.dumps(data or {})
         app.redis_client.set(cls._get_redis_extra_data_key(type_, user_id), json_data, ex=ttl)
         token = cls.load_without_checking(encoded_token, type_, user_id)
         token._log(cls._TokenAction.CREATE)
         return token
 
-    def expire(self) -> None:
-        super().expire()
-        app.redis_client.delete(self._get_redis_extra_data_key(self.type_, self.user_id))
-
     @classmethod
-    def delete(cls, type_: TokenType, user_id: int) -> None:
+    def delete(cls, type_: TokenType, user_id: int | str | None) -> None:
+        if not isinstance(user_id, int):
+            raise ValueError("user_id can only be an int for SixDigitsToken")
         super().delete(type_, user_id)
         app.redis_client.delete(cls._get_redis_extra_data_key(type_, user_id))
 
-    @classmethod
-    def get_token(cls, type_: TokenType, user_id: int) -> "SixDigitsToken | None":
-        encoded_token = app.redis_client.get(cls._get_redis_key(type_, user_id))
-        if encoded_token is None:
-            return None
-        return cls.load_without_checking(encoded_token, type_, user_id)
+    def expire(self) -> None:
+        assert isinstance(self.key_suffix, int), "SixDigitsToken key suffix can only be a user id"
+        super().expire()
+        app.redis_client.delete(self._get_redis_extra_data_key(self.type_, self.key_suffix))
